Pedro/logging (#2)
* removes printing and adds logging

* Closes #12
