--- gmnspy/conversion.py
@@ -1,2 +1,0 @@
-def osmnx_to_gmns():
-    pass
